pyproject.toml
[tool.poetry]
name = "analyse-interviews"
version = "0.1.0"
description = "Analyse qualitative d'interviews à l'aide de gpt"
authors = ["opmvpc <thibsix@outlook.be>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.40.1"
pandas = "^2.2.2"
python-docx = "^1.1.2"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
rich = "^13.7.1"
ollama = "^0.3.1"
prettytable = "^3.10.2"
matplotlib = "^3.9.1.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
